Lock screen in specific case
