+++ app/src/main/kotlin/com/saltedge/authenticator/app/Constants.kt
@@ -40,4 +40,4 @@ const val KEY_CONNECT_CONFIGURATION = "KEY_CONNECT_CONFIGURATION"
 
 const val TIME_VIEW_UPDATE_TIMEOUT = 1000L
 
-const val KEY_SKIP_PIN = "KEY_SKIP_PIN"
+const val LOCK_ON_CREATE = "LOCK_ON_CREATE"
