+++ app/src/main/kotlin/com/saltedge/authenticator/features/launcher/LauncherActivity.kt
@@ -27,6 +27,7 @@ import android.os.Handler
 import androidx.appcompat.app.AppCompatActivity
 import com.saltedge.authenticator.R
 import com.saltedge.authenticator.app.KEY_DEEP_LINK
+import com.saltedge.authenticator.app.LOCK_ON_CREATE
 import com.saltedge.authenticator.features.launcher.di.LauncherModule
 import com.saltedge.authenticator.model.realm.RealmManager
 import com.saltedge.authenticator.sdk.constants.KEY_AUTHORIZATION_ID
@@ -70,6 +71,7 @@ class LauncherActivity : AppCompatActivity() {
             .putExtra(KEY_CONNECTION_ID, intent.getStringExtra(KEY_CONNECTION_ID))
             .putExtra(KEY_AUTHORIZATION_ID, intent.getStringExtra(KEY_AUTHORIZATION_ID))
             .putExtra(KEY_DEEP_LINK, intent.dataString)
+            .putExtra(LOCK_ON_CREATE, true)
             .apply {
                 flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK
             })
